components/form: refactor types
