+++ src/components/types/Form.ts
@@ -1,0 +1,9 @@
+export type FormPersonalDetailsFields = {
+  firstName: string;
+  lastName: string;
+  email: string;
+  nickname: string;
+  gender: string;
+  newsletter: string[];
+  terms: boolean;
+};
